gatgu/gatgu/urls.py
from django.conf import settings
from django.conf.urls import url
from django.contrib import admin
from django.urls import path, include

urlpatterns = [
    path('admin/', admin.site.urls),
    
    path('v1/', include('user.urls')),
    path('v1/', include('article.urls')),
    path('v1/chat/', include('chat.urls'))

]

if settings.DEBUG_TOOLBAR:
    import debug_toolbar

    urlpatterns += [
        url(r'^__debug__/', include(debug_toolbar.urls)),
    ]