requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
autobahn==21.2.2
Automat==20.2.0
backends==0.5.4
boto3==1.17.8
botocore==1.20.8
certifi==2020.11.8
cffi==1.14.5
channels==3.0.3
chardet==3.0.4
constantly==15.1.0
coverage==5.3
cryptography==3.4.6
daphne==3.0.1
Django==3.1
django-debug-toolbar==2.2
django-redis==4.12.1
django-storages==1.11.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
fdm==0.4.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mpmath==1.2.1
mysqlclient==2.0.1
Naked==0.1.31
numpy==1.20.1
Pillow==8.1.0
plum-dispatch==0.2.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.0
s3transfer==0.3.4
scipy==1.6.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
status==0.2.5
sympy==1.7.1
toml==0.10.2
Twisted==20.3.0
txaio==21.2.1

urllib3==1.26.2
wrapt==1.12.1
zope.interface==5.2.0
